script: Move performance DOM interfaces to `script/dom/performance/`
