+++ components/script/dom/globalscope.rs
@@ -117,8 +117,8 @@ use crate::dom::html::htmlscriptelement::{ScriptId, SourceCode};
 use crate::dom::htmlscriptelement::ScriptOrigin;
 use crate::dom::messageport::MessagePort;
 use crate::dom::paintworkletglobalscope::PaintWorkletGlobalScope;
-use crate::dom::performance::Performance;
-use crate::dom::performanceobserver::VALID_ENTRY_TYPES;
+use crate::dom::performance::performance::Performance;
+use crate::dom::performance::performanceobserver::VALID_ENTRY_TYPES;
 use crate::dom::promise::Promise;
 use crate::dom::readablestream::{CrossRealmTransformReadable, ReadableStream};
 use crate::dom::reportingobserver::ReportingObserver;
