+++ components/script/dom/html/htmlimageelement.rs
@@ -75,7 +75,7 @@ use crate::dom::html::htmlsourceelement::HTMLSourceElement;
 use crate::dom::medialist::MediaList;
 use crate::dom::mouseevent::MouseEvent;
 use crate::dom::node::{BindContext, Node, NodeDamage, NodeTraits, ShadowIncluding, UnbindContext};
-use crate::dom::performanceresourcetiming::InitiatorType;
+use crate::dom::performance::performanceresourcetiming::InitiatorType;
 use crate::dom::promise::Promise;
 use crate::dom::virtualmethods::VirtualMethods;
 use crate::dom::window::Window;
