+++ components/script/dom/html/htmllinkelement.rs
@@ -51,7 +51,7 @@ use crate::dom::element::{
 use crate::dom::html::htmlelement::HTMLElement;
 use crate::dom::medialist::MediaList;
 use crate::dom::node::{BindContext, Node, NodeTraits, UnbindContext};
-use crate::dom::performanceresourcetiming::InitiatorType;
+use crate::dom::performance::performanceresourcetiming::InitiatorType;
 use crate::dom::processingoptions::{
     LinkFetchContext, LinkFetchContextType, LinkProcessingOptions,
 };
