+++ components/script/dom/performance/performancepainttiming.rs
@@ -7,11 +7,11 @@ use dom_struct::dom_struct;
 use script_traits::ProgressiveWebMetricType;
 use time::Duration;
 
+use super::performanceentry::PerformanceEntry;
 use crate::dom::bindings::reflector::reflect_dom_object;
 use crate::dom::bindings::root::DomRoot;
 use crate::dom::bindings::str::DOMString;
 use crate::dom::globalscope::GlobalScope;
-use crate::dom::performanceentry::PerformanceEntry;
 use crate::script_runtime::CanGc;
 
 #[dom_struct]
