+++ components/script/dom/workers/workerglobalscope.rs
@@ -71,8 +71,8 @@ use crate::dom::dedicatedworkerglobalscope::{
 use crate::dom::globalscope::GlobalScope;
 use crate::dom::htmlscriptelement::{SCRIPT_JS_MIMES, ScriptOrigin, ScriptType};
 use crate::dom::indexeddb::idbfactory::IDBFactory;
-use crate::dom::performance::Performance;
-use crate::dom::performanceresourcetiming::InitiatorType;
+use crate::dom::performance::performance::Performance;
+use crate::dom::performance::performanceresourcetiming::InitiatorType;
 use crate::dom::promise::Promise;
 use crate::dom::reportingendpoint::{ReportingEndpoint, SendReportsToEndpoints};
 use crate::dom::reportingobserver::ReportingObserver;
